+++ e2e/specs/st_hello.spec.ts
@@ -19,6 +19,8 @@
 
 describe("hello", () => {
   before(() => {
+    // Increasing timeout since we're waiting for the animation and map to load.
+    Cypress.config("defaultCommandTimeout", 30000);
     cy.visit("http://localhost:3000/");
   });
 
@@ -29,5 +31,106 @@ describe("hello", () => {
     );
 
     cy.get(".streamlit-dialog").should("not.exist");
+
+    cy.get(".stSelectbox").should("exist");
+
+    cy.get(".reportview-container").matchImageSnapshot("welcome-streamlit");
+  });
+
+  it("displays animation demo", () => {
+    cy.get(".element-container .stSelectbox")
+      .click()
+      .then(() => {
+        cy.get("ul li:nth-child(2)")
+          .last()
+          .click()
+          .then(() => {
+            cy.get(".element-container .stMarkdown h1").should(
+              "contain",
+              "Animation Demo"
+            );
+
+            // Wait for the animation to end.
+            cy.get(".stButton button").contains("Re-run");
+
+            cy.get(".reportview-container").matchImageSnapshot(
+              "animation-demo"
+            );
+          });
+      });
+  });
+
+  it("displays plotting demo", () => {
+    cy.get(".element-container .stSelectbox")
+      .click()
+      .then(() => {
+        cy.get("ul li:nth-child(3)")
+          .last()
+          .click()
+          .then(() => {
+            cy.get(".element-container .stMarkdown h1").should(
+              "contain",
+              "Plotting Demo"
+            );
+
+            // Wait for the animation to end.
+            cy.get(".stText").contains("100% Complete");
+
+            cy.get(".element-container .stVegaLiteChart")
+              .find("canvas")
+              .should("have.css", "height", "300px");
+          });
+      });
+  });
+
+  it("displays mapping demo", () => {
+    cy.get(".element-container .stSelectbox")
+      .click()
+      .then(() => {
+        cy.get("ul li:nth-child(4)")
+          .click()
+          .then(() => {
+            cy.get(".element-container .stMarkdown h1").should(
+              "contain",
+              "Mapping Demo"
+            );
+
+            cy.get(".element-container .stDeckGlJsonChart")
+              .find("canvas")
+              .should("have.css", "height", "500px");
+
+            // Wait for Mapbox to build the canvas.
+            cy.wait(5000);
+
+            cy.get(".reportview-container").matchImageSnapshot("mapping-demo");
+          });
+      });
+  });
+
+  it("displays dataframe demo", () => {
+    cy.get(".element-container .stSelectbox")
+      .click()
+      .then(() => {
+        cy.get("ul li:nth-child(5)")
+          .click()
+          .then(() => {
+            cy.get(".element-container .stMarkdown h1").should(
+              "contain",
+              "DataFrame Demo"
+            );
+
+            cy.get(".stMultiSelect").should("exist");
+
+            cy.get(".stDataFrame").should("exist");
+
+            cy.get(".element-container .stVegaLiteChart")
+              .find("canvas")
+              .should("have.css", "height", "300px");
+
+            cy.get(".reportview-container").matchImageSnapshot(
+              "dataframe-demo"
+            );
+          });
+      });
   });
 });