+++ lib/tests/streamlit/components_test.py
@@ -15,6 +15,7 @@
 import json
 import os
 import unittest
+from typing import Any
 from unittest import mock
 
 import pandas as pd
@@ -28,12 +29,29 @@ from streamlit.components.v1.components import ComponentRequestHandler
 from streamlit.components.v1.components import CustomComponent
 from streamlit.components.v1.components import declare_component
 import streamlit.components.v1 as components
+from streamlit.elements import arrow_table
 from streamlit.errors import DuplicateWidgetID
+from streamlit.proto.ComponentInstance_pb2 import SpecialArg
+from streamlit.type_util import to_bytes
 from tests import testutil
 from tests.testutil import DeltaGeneratorTestCase
 
 URL = "http://not.a.real.url:3001"
 PATH = "not/a/real/path"
+
+
+def _serialize_dataframe_arg(key: str, value: Any) -> SpecialArg:
+    special_arg = SpecialArg()
+    special_arg.key = key
+    arrow_table.marshall(special_arg.arrow_dataframe.data, value)
+    return special_arg
+
+
+def _serialize_bytes_arg(key: str, value: Any) -> SpecialArg:
+    special_arg = SpecialArg()
+    special_arg.key = key
+    special_arg.bytes = to_bytes(value)
+    return special_arg
 
 
 class DeclareComponentTest(unittest.TestCase):
@@ -205,8 +223,8 @@ class InvokeComponentTest(DeltaGeneratorTestCase):
         proto = self.get_delta_from_queue().new_element.component_instance
 
         self.assertEqual(self.test_component.name, proto.component_name)
-        self.assertEqual(json.dumps({"foo": "bar"}), proto.args_json)
-        self.assertEqual("[]", str(proto.args_dataframe))
+        self.assertEqual(json.dumps({"foo": "bar"}), proto.json_args)
+        self.assertEqual("[]", str(proto.special_args))
 
     def test_only_df_args(self):
         """Test that component with only dataframe args is marshalled correctly."""
@@ -220,16 +238,16 @@ class InvokeComponentTest(DeltaGeneratorTestCase):
         proto = self.get_delta_from_queue().new_element.component_instance
 
         self.assertEqual(self.test_component.name, proto.component_name)
-        self.assertEqual("{}", proto.args_json)
-        # (HK) TODO: Add assertEqual check for Apache Arrow pybytes.
-        self.assertIsNotNone(proto.args_dataframe)
+        self.assertEqual("{}", proto.json_args)
+        self.assertEqual(1, len(proto.special_args))
+        self.assertEqual(_serialize_dataframe_arg("df", df), proto.special_args[0])
 
     def test_only_list_args(self):
         """Test that component with only list args is marshalled correctly."""
         self.test_component(data=["foo", "bar", "baz"])
         proto = self.get_delta_from_queue().new_element.component_instance
-        self.assertEqual(json.dumps({"data": ["foo", "bar", "baz"]}), proto.args_json)
-        self.assertEqual("[]", str(proto.args_dataframe))
+        self.assertEqual(json.dumps({"data": ["foo", "bar", "baz"]}), proto.json_args)
+        self.assertEqual("[]", str(proto.special_args))
 
     def test_no_args(self):
         """Test that component with no args is marshalled correctly."""
@@ -237,26 +255,41 @@ class InvokeComponentTest(DeltaGeneratorTestCase):
         proto = self.get_delta_from_queue().new_element.component_instance
 
         self.assertEqual(self.test_component.name, proto.component_name)
-        self.assertEqual("{}", proto.args_json)
-        self.assertEqual("[]", str(proto.args_dataframe))
+        self.assertEqual("{}", proto.json_args)
+        self.assertEqual("[]", str(proto.special_args))
 
-    def test_json_and_df_args(self):
-        """Test that component with json and dataframe args is marshalled correctly."""
-        raw_data = {
-            "First Name": ["Jason", "Molly"],
-            "Last Name": ["Miller", "Jacobson"],
-            "Age": [42, 52],
-        }
-        df = pd.DataFrame(raw_data, columns=["First Name", "Last Name", "Age"])
-        self.test_component(foo="bar", df=df)
+    def test_bytes_args(self):
+        self.test_component(foo=b"foo", bar=bytearray(b"bar"))
+        proto = self.get_delta_from_queue().new_element.component_instance
+        self.assertEqual(json.dumps({}), proto.json_args)
+        self.assertEqual(2, len(proto.special_args))
+        self.assertEqual(_serialize_bytes_arg("foo", b"foo"), proto.special_args[0])
+        self.assertEqual(
+            _serialize_bytes_arg("bar", bytearray(b"bar")), proto.special_args[1]
+        )
+
+    def test_mixed_args(self):
+        """Test marshalling of a component with varied arg types."""
+        df = pd.DataFrame(
+            {
+                "First Name": ["Jason", "Molly"],
+                "Last Name": ["Miller", "Jacobson"],
+                "Age": [42, 52],
+            },
+            columns=["First Name", "Last Name", "Age"],
+        )
+        self.test_component(string_arg="string", df_arg=df, bytes_arg=b"bytes")
         proto = self.get_delta_from_queue().new_element.component_instance
 
         self.assertEqual(self.test_component.name, proto.component_name)
-        self.assertEqual(json.dumps({"foo": "bar"}), proto.args_json)
-        # (HK) TODO: Add assertEqual check for Apache Arrow pybytes.
-        self.assertIsNotNone(proto.args_dataframe)
+        self.assertEqual(json.dumps({"string_arg": "string"}), proto.json_args)
+        self.assertEqual(2, len(proto.special_args))
+        self.assertEqual(_serialize_dataframe_arg("df_arg", df), proto.special_args[0])
+        self.assertEqual(
+            _serialize_bytes_arg("bytes_arg", b"bytes"), proto.special_args[1]
+        )
 
-    def test_key(self):
+    def test_duplicate_key(self):
         """Two components with the same `key` should throw DuplicateWidgetID exception"""
         self.test_component(foo="bar", key="baz")